+++ lib/services/common/channel.dart
@@ -1,5 +1,3 @@
-import 'dart:isolate';
-
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 import 'package:streams_channel/streams_channel.dart';
@@ -12,7 +10,7 @@ class AvesMethodChannel extends MethodChannel {
   @override
   Future<T?> invokeMethod<T>(String method, [arguments]) {
     if (kDebug) {
-      debugPrint('$runtimeType platform call isolate=${Isolate.current.debugName} channel=$name method=$method arguments=$arguments');
+      debugPrint('$runtimeType platform call channel=$name method=$method arguments=$arguments');
     }
     return super.invokeMethod(method, arguments);
   }
@@ -24,7 +22,7 @@ class AvesStreamsChannel extends StreamsChannel {
   @override
   Stream receiveBroadcastStream([arguments]) {
     if (AvesMethodChannel.kDebug) {
-      debugPrint('$runtimeType platform call isolate=${Isolate.current.debugName} channel=$name arguments=$arguments');
+      debugPrint('$runtimeType platform call channel=$name arguments=$arguments');
     }
     return super.receiveBroadcastStream(arguments);
   }