+++ lib/widgets/aves_app.dart
@@ -24,7 +24,9 @@ import 'package:aves/services/common/services.dart';
 import 'package:aves/theme/colors.dart';
 import 'package:aves/theme/icons.dart';
 import 'package:aves/theme/styles.dart';
+import 'package:aves/theme/text.dart';
 import 'package:aves/theme/themes.dart';
+import 'package:aves/widgets/about/app_ref.dart';
 import 'package:aves/widgets/collection/collection_grid.dart';
 import 'package:aves/widgets/collection/collection_page.dart';
 import 'package:aves/widgets/common/action_mixins/feedback.dart';
@@ -294,7 +296,10 @@ class _AvesAppState extends State<AvesApp> with WidgetsBindingObserver {
                                 navigatorKey: navigatorKey,
                                 home: home,
                                 onUnknownRoute: (settings) {
-                                  reportService.recordError(Exception('Could not find a generator for route $settings in the $runtimeType.'));
+                                  // as of Flutter v3.44.2, using `$settings` in exception message yields `Instance of 'RouteSettings'` in reports,
+                                  // so we explicitly stringify variable outside
+                                  final settingsString = '${settings.runtimeType}(${settings.name == null ? 'none' : '"${settings.name}"'}, ${settings.arguments})';
+                                  reportService.recordError(Exception('Could not find a generator for route settings=$settingsString in the $runtimeType.'));
                                   return null;
                                 },
                                 navigatorObservers: _navigatorObservers,
@@ -767,7 +772,16 @@ class _AvesAppContentDecoratorState extends State<AvesAppContentDecorator> with 
             if (success) {
               await reportService.log('Exported settings to dirPath=$dirPath');
             } else {
-              showFeedback(context, FeedbackType.warn, context.l10n.genericFailureFeedback);
+              final l10n = context.l10n;
+              showFeedback(
+                context,
+                FeedbackType.warn,
+                '${l10n.genericFailureFeedback}${AText.separator}${l10n.settingsAutoExportSettings}',
+                SnackBarAction(
+                  label: 'FAQ',
+                  onPressed: () => AvesApp.launchUrl('${AppReference.avesFaq}#why-is-auto-settings-export-failing'),
+                ),
+              );
             }
           }
         }